moved everything into a base directory for the actual app
