+++ app/qrgen.py
